requirement.txt
tensorflow
ultralytics
opencv-python
numpy
pandas
matplotlib
scikit-learn
streamlit
Pillow
datasets